Add Answer abstraction layer (#174)
